--- src/components/Sidebar/Item/index.module.css
@@ -1,58 +1,0 @@
-.sub_menu {
-  background-color: var(--white-transparent);
-  border-radius: 5px;
-}
-
-.sub_menu .menu_item {
-  padding: 10px;
-}
-
-.sub_menu_title {
-  display: flex;
-  justify-content: space-between;
-}
-
-.menu_item {
-  padding: 10px 0;
-}
-
-.menu_icon {
-  width: 14px;
-}
-
-.menu_item a,
-.sub_menu_title {
-  color: var(--white-0);
-  text-decoration: none;
-  padding: 4px 0;
-  cursor: pointer;
-}
-
-.menu_item:hover a,
-.menu_item:hover svg,
-.sub_menu_title:hover,
-.active {
-  color: var(--orange-2);
-}
-
-.bottom {
-  border-top: 1px solid var(--blue-1);
-  position: absolute;
-  bottom: 20px;
-  background-color: var(--blue-0);
-  width: calc(100% - 40px);
-}
-
-.bottom .menu_item {
-  margin: 10px 0 40px;
-}
-
-.menu_item svg {
-  color: var(--white-0);
-  margin: 0 10px 0 0;
-}
-
-.button {
-  width: 100%;
-  margin: 4px 0;
-}
