+++ src/components/Sidebar/index.module.css
@@ -10,13 +10,70 @@
   position: sticky;
 }
 
-.menu {
+.items {
   color: var(--white-0);
   margin: 30px 0;
   max-height: 70vh;
   overflow-y: auto;
 }
 
+.sub {
+  background-color: var(--white-transparent);
+  border-radius: 5px;
+}
+
+.sub .item {
+  padding: 10px;
+}
+
+.sub_title {
+  display: flex;
+  justify-content: space-between;
+}
+
+.item {
+  padding: 10px 0;
+}
+
+.item svg {
+  color: var(--white-0);
+  margin: 0 10px 0 0;
+  width: 14px;
+}
+
+.item a,
+.sub_title {
+  color: var(--white-0);
+  text-decoration: none;
+  padding: 4px 0;
+  cursor: pointer;
+}
+
+.item:hover a,
+.item:hover svg,
+.sub_title:hover,
+.active a,
+.active a svg {
+  color: var(--orange-2);
+}
+
 .logo {
   height: 40px;
+}
+
+.bottom {
+  border-top: 1px solid var(--blue-1);
+  position: absolute;
+  bottom: 20px;
+  background-color: var(--blue-0);
+  width: calc(100% - 40px);
+}
+
+.bottom .item {
+  margin: 10px 0 40px;
+}
+
+.bottom .button {
+  width: 100%;
+  margin: 4px 0;
 }
